requirements.txt
anyio==3.6.1
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.0
cryptography==37.0.2
dadata==21.10.1
dj-database-url==0.5.0
Django==4.0.5
django-etc==1.3.1
django-extensions==3.1.5
django-sitemessage==1.3.4
djangorestframework==3.13.1
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
pycparser==2.21
pyOpenSSL==22.0.0
pyTelegramBotAPI==4.6.0
pytz==2022.1
requests==2.28.0
rfc3986==1.5.0
schedule==1.1.0
sniffio==1.2.0
sqlparse==0.4.2
urllib3==1.26.9
Werkzeug==2.1.2
whitenoise==6.2.0
xlwt==1.3.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django_spurl==0.6.8
# psycopg2==2.9.3
